refactor(kitchenpos): merge menu's BC and menugroup's BC
