--- src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,20 +1,0 @@
-package kitchenpos.menu.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-import kitchenpos.menu.domain.Menu;
-
-public interface MenuRepository {
-    Menu save(Menu menu);
-
-    Optional<Menu> findById(UUID id);
-
-    List<Menu> findAll();
-
-    List<Menu> findAllByIdIn(List<UUID> ids);
-
-    List<Menu> findAllByProductId(UUID productId);
-}
-
