+++ src/main/java/kitchenpos/menu/domain/menu/MenuRepository.java
@@ -1,0 +1,18 @@
+package kitchenpos.menu.domain.menu;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface MenuRepository {
+    Menu save(Menu menu);
+
+    Optional<Menu> findById(UUID id);
+
+    List<Menu> findAll();
+
+    List<Menu> findAllByIdIn(List<UUID> ids);
+
+    List<Menu> findAllByProductId(UUID productId);
+}
+
