+++ src/main/java/kitchenpos/menu/domain/menugroup/JpaMenuGroupRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.menu.domain.menugroup;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
+}
