+++ src/main/java/kitchenpos/menu/domain/menugroup/MenuGroup.java
@@ -1,0 +1,37 @@
+package kitchenpos.menu.domain.menugroup;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.util.UUID;
+
+@Table(name = "menu_group")
+@Entity
+public class MenuGroup {
+    @Column(name = "id", columnDefinition = "varbinary(16)")
+    @Id
+    private UUID id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    public MenuGroup() {
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(final UUID id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+}
