--- src/main/java/kitchenpos/menugroup/domain/JpaMenuGroupRepository.java
@@ -1,11 +1,0 @@
-package kitchenpos.menugroup.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
-
-public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
-}
