--- src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,14 +1,0 @@
-package kitchenpos.menugroup.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface MenuGroupRepository {
-    MenuGroup save(MenuGroup menuGroup);
-
-    Optional<MenuGroup> findById(UUID id);
-
-    List<MenuGroup> findAll();
-}
-
