+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.application;
 
 import kitchenpos.order.domain.KitchenridersClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menu.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
