+++ src/test/java/kitchenpos/menu/domain/menugroup/InMemoryMenuGroupRepository.java
@@ -1,0 +1,26 @@
+package kitchenpos.menu.domain.menugroup;
+
+import kitchenpos.menu.domain.menugroup.MenuGroup;
+import kitchenpos.menu.domain.menugroup.MenuGroupRepository;
+
+import java.util.*;
+
+public class InMemoryMenuGroupRepository implements MenuGroupRepository {
+    private final Map<UUID, MenuGroup> menuGroups = new HashMap<>();
+
+    @Override
+    public MenuGroup save(final MenuGroup menuGroup) {
+        menuGroups.put(menuGroup.getId(), menuGroup);
+        return menuGroup;
+    }
+
+    @Override
+    public Optional<MenuGroup> findById(final UUID id) {
+        return Optional.ofNullable(menuGroups.get(id));
+    }
+
+    @Override
+    public List<MenuGroup> findAll() {
+        return new ArrayList<>(menuGroups.values());
+    }
+}
