--- src/test/java/kitchenpos/menu/infra/repository/InMemoryMenuRepository.java
@@ -1,43 +1,0 @@
-package kitchenpos.menu.infra.repository;
-
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-public class InMemoryMenuRepository implements MenuRepository {
-    private final Map<UUID, Menu> menus = new HashMap<>();
-
-    @Override
-    public Menu save(final Menu menu) {
-        menus.put(menu.getId(), menu);
-        return menu;
-    }
-
-    @Override
-    public Optional<Menu> findById(final UUID id) {
-        return Optional.ofNullable(menus.get(id));
-    }
-
-    @Override
-    public List<Menu> findAll() {
-        return new ArrayList<>(menus.values());
-    }
-
-    @Override
-    public List<Menu> findAllByIdIn(final List<UUID> ids) {
-        return menus.values()
-            .stream()
-            .filter(menu -> ids.contains(menu.getId()))
-            .collect(Collectors.toList());
-    }
-
-    @Override
-    public List<Menu> findAllByProductId(final UUID productId) {
-        return menus.values()
-            .stream()
-            .filter(menu -> menu.getMenuProducts().stream().anyMatch(menuProduct -> menuProduct.getProduct().getId().equals(productId)))
-            .collect(Collectors.toList());
-    }
-}
