--- src/test/java/kitchenpos/menugroup/infra/repository/InMemoryMenuGroupRepository.java
@@ -1,26 +1,0 @@
-package kitchenpos.menugroup.infra.repository;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
-
-import java.util.*;
-
-public class InMemoryMenuGroupRepository implements MenuGroupRepository {
-    private final Map<UUID, MenuGroup> menuGroups = new HashMap<>();
-
-    @Override
-    public MenuGroup save(final MenuGroup menuGroup) {
-        menuGroups.put(menuGroup.getId(), menuGroup);
-        return menuGroup;
-    }
-
-    @Override
-    public Optional<MenuGroup> findById(final UUID id) {
-        return Optional.ofNullable(menuGroups.get(id));
-    }
-
-    @Override
-    public List<MenuGroup> findAll() {
-        return new ArrayList<>(menuGroups.values());
-    }
-}
