+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.order.application;
 
 import kitchenpos.order.infra.client.FakeKitchenridersClient;
-import kitchenpos.menu.infra.repository.InMemoryMenuRepository;
-import kitchenpos.order.infra.repository.InMemoryOrderRepository;
-import kitchenpos.ordertable.infra.repository.InMemoryOrderTableRepository;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.menu.InMemoryMenuRepository;
+import kitchenpos.order.domain.InMemoryOrderRepository;
+import kitchenpos.ordertable.domain.InMemoryOrderTableRepository;
+import kitchenpos.menu.domain.menu.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
@@ -17,7 +17,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.*;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.junit.jupiter.params.provider.NullSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.math.BigDecimal;
 import java.util.*;
