+++ src/test/java/kitchenpos/order/domain/InMemoryOrderRepository.java
@@ -1,0 +1,35 @@
+package kitchenpos.order.domain;
+
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
+
+import java.util.*;
+
+public class InMemoryOrderRepository implements OrderRepository {
+    private final Map<UUID, Order> orders = new HashMap<>();
+
+    @Override
+    public Order save(final Order order) {
+        orders.put(order.getId(), order);
+        return order;
+    }
+
+    @Override
+    public Optional<Order> findById(final UUID id) {
+        return Optional.ofNullable(orders.get(id));
+    }
+
+    @Override
+    public List<Order> findAll() {
+        return new ArrayList<>(orders.values());
+    }
+
+    @Override
+    public boolean existsByOrderTableAndStatusNot(final OrderTable orderTable, final OrderStatus status) {
+        return orders.values()
+            .stream()
+            .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
+    }
+}
